test(js-sdk): move deploy suite to tests/runtimes/cloudflare-deploy
Sibling directories instead of nesting: tests/runtimes/cloudflare is
the vitest-pool-workers suite, tests/runtimes/cloudflare-deploy the
real-infrastructure one (suite name cloudflare-deploy).
